Refactored App.jsx
