--- frontend/src/App.css
@@ -1,14 +1,0 @@
-.App {
-  text-align: center;
-}
-
-.App-header {
-  background-color: #EDF4ED;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
-  font-family: 'Roboto', sans-serif;
-}
