+++ frontend/src/components/AssessmentCard.js
@@ -7,7 +7,7 @@ import { Divider } from '@mui/material';
 import Questioner from './Questioner';
 
 const AssessmentCard = ({ details, counter, setCounter, results, setResults }) => {
-  
+
   useEffect(() => {
     console.log(results);
   }, [results]);
@@ -27,7 +27,7 @@ const AssessmentCard = ({ details, counter, setCounter, results, setResults }) =
 
   return (
     <Box flex sx={{ padding: 2, backgroundColor: '#FFFFFF', borderRadius: 1, width: '40%', fontSize: '1.1rem' }}>
-      <Stack spacing={3}>
+      <Stack spacing={3} justifyContent={'space-evenly'}>
         <p style={{ textAlign: 'left' }}> Assessment: {details.content.display_name} </p>
         <p style={{ textAlign: 'left' }}> {section.title} </p>
         <Divider textAlign='right' border='1px solid'> {counter + 1} out of {questions.length} </Divider>
@@ -36,7 +36,7 @@ const AssessmentCard = ({ details, counter, setCounter, results, setResults }) =
           answers={answers}
           handleClick={handleAnswerClick}
         />
-        <LinearProgress variant="determinate" value={(counter)/questions.length * 100} />
+        <LinearProgress variant="determinate" value={(counter)/questions.length * 100} sx={{height: '10px', borderRadius: 3}}/>
       </Stack>
     </Box>
   );
